requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
attrs==25.4.0
Django==5.1
djangorestframework==3.15.0
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.11.0
gunicorn==22.0.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
pydantic==2.9.0
pydantic_core==2.23.2
pytest==8.3.0
pytest-django==4.9.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.27.1
ruff==0.7.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
